# dict/api/appointment/index.ts
/* eslint-disable */
// tslint:disable
/**
 * Patients Appointment
 * Patient Appointment
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import passport from 'passport'
import * as t from './types'
import * as v from '../../validation'
import { Api } from '../../models'

export default function(app: Express, impl: t.AppointmentApi) {
	app.post(
		'/appointment/create',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiAppointmentDtFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.postAppointmentCreate(__body()).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiAppointmentDtToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.postAppointmentCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 401) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.postAppointmentCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(401)
						res.send(body)
						return
					}
					if (response.status === 404) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.postAppointmentCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(404)
						res.send(body)
						return
					}
					if (response.status === 422) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.postAppointmentCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(422)
						res.send(body)
						return
					}

					console.log('Unsupported response in appointment.postAppointmentCreate', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in appointment.postAppointmentCreate', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.delete(
		'/appointment/delete',
		function (req, res) {
			try {
				impl.deleteAppointmentDelete(v.parseString('query.Id', req.query['Id'])).then(function (response) {
					if (response.status === 200) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.deleteAppointmentDelete', error)
							res.status(500)
							res.send()
							return
						}

						res.status(200)
						res.send(body)
						return
					}
					if (response.status === 401) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.deleteAppointmentDelete', error)
							res.status(500)
							res.send()
							return
						}

						res.status(401)
						res.send(body)
						return
					}
					if (response.status === 404) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.deleteAppointmentDelete', error)
							res.status(500)
							res.send()
							return
						}

						res.status(404)
						res.send(body)
						return
					}
					if (response.status === 422) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.deleteAppointmentDelete', error)
							res.status(500)
							res.send()
							return
						}

						res.status(422)
						res.send(body)
						return
					}

					console.log('Unsupported response in appointment.deleteAppointmentDelete', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in appointment.deleteAppointmentDelete', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.get(
		'/appointment/get',
		function (req, res) {
			try {
				impl.getAppointmentGet(v.parseString('query.Id', req.query['Id'])).then(function (response) {
					if (response.status === 200) {
						let body: any
						try {
							body = v.modelApiAppointmentDtToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGet', error)
							res.status(500)
							res.send()
							return
						}

						res.status(200)
						res.send(body)
						return
					}
					if (response.status === 401) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGet', error)
							res.status(500)
							res.send()
							return
						}

						res.status(401)
						res.send(body)
						return
					}
					if (response.status === 404) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGet', error)
							res.status(500)
							res.send()
							return
						}

						res.status(404)
						res.send(body)
						return
					}
					if (response.status === 422) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGet', error)
							res.status(500)
							res.send()
							return
						}

						res.status(422)
						res.send(body)
						return
					}

					console.log('Unsupported response in appointment.getAppointmentGet', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in appointment.getAppointmentGet', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.get(
		'/appointment/getAll',
		function (req, res) {
			try {
				impl.getAppointmentGetAll(v.allowUndefined(v.parseInteger)('query.Limit', req.query['Limit']), v.allowUndefined(v.enumApiDirectionParamEnumFromJson)('query.Direction', req.query['Direction']), v.allowUndefined(v.parseString)('query.SortByField', req.query['SortByField'])).then(function (response) {
					if (response.status === 200) {
						let body: any
						try {
							body = v.modelApiAppointmentPagedResultDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGetAll', error)
							res.status(500)
							res.send()
							return
						}

						res.status(200)
						res.send(body)
						return
					}
					if (response.status === 401) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGetAll', error)
							res.status(500)
							res.send()
							return
						}

						res.status(401)
						res.send(body)
						return
					}
					if (response.status === 404) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.getAppointmentGetAll', error)
							res.status(500)
							res.send()
							return
						}

						res.status(404)
						res.send(body)
						return
					}

					console.log('Unsupported response in appointment.getAppointmentGetAll', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in appointment.getAppointmentGetAll', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.put(
		'/appointment/update/:id',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiAppointmentDtFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.putAppointmentUpdate(v.allowUndefined(v.parseString)('query.id', req.query['id']), __body()).then(function (response) {
					if (response.status === 200) {
						let body: any
						try {
							body = v.modelApiAppointmentDtToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.putAppointmentUpdate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(200)
						res.send(body)
						return
					}
					if (response.status === 401) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.putAppointmentUpdate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(401)
						res.send(body)
						return
					}
					if (response.status === 404) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.putAppointmentUpdate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(404)
						res.send(body)
						return
					}
					if (response.status === 422) {
						let body: any
						try {
							body = v.modelApiMessageDtoToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in appointment.putAppointmentUpdate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(422)
						res.send(body)
						return
					}

					console.log('Unsupported response in appointment.putAppointmentUpdate', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in appointment.putAppointmentUpdate', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

}