# dict/models.ts
/* eslint-disable */
// tslint:disable
/**
 * Patients Appointment
 * Patient Appointment
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

export namespace Api {
	export interface AppointmentDt {
		id?: string;
		patientid: string;
		doctorid: string;
		doctorName?: string;
		doctorImage?: string;
		createdAt?: string;
		symptoms: string;
		slotid: string;
		slotTime: string;
		appointmentDate: string;
		/**
		 * @type {number}
		 * @memberof AppointmentDt
		 */
		appointmentType: number;
		appointmentStatus?: string;
	}

	export interface AppointmentPagedResultDto {
		/**
		 * @type {number}
		 * @memberof AppointmentPagedResultDto
		 */
		totalCount?: number;
		items?: Api.AppointmentDt[];
	}

	export enum DirectionParamEnum {
		Asc = 'asc',
		Desc = 'desc'
	}

	export interface MessageDto {
		message: string;
	}

}